changed mutateChange to a double

+++ Cuda/Genetic_Algorithm/ga_crossover.cpp
@@ -233,8 +233,8 @@ rkParameters<double> mutate(const rkParameters<double> & p1, std::mt19937_64 & r
     rkParameters<double> newInd = p1;
 
     int genesToMutate = 1; // number of genes to mutate
-    int mutateChance = (static_cast<double>(rng()) / rng.max());
-
+    double mutateChance = (static_cast<double>(rng()) / rng.max());
+    
     if (mutateChance < cConstants->triple_mutation_rate) {
         genesToMutate = 3;
     }
